requirements.txt
einops==0.7.0
faiss_gpu==1.7.2
fast_pytorch_kmeans==0.2.0.1
geopy==2.4.0
numpy==1.26.1
opencv_python==4.8.1.78
Pillow==10.1.0
pytorch_lightning==2.1.0
Requests==2.31.0
scikit_learn==1.3.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
